Computer vision/requirements_yolo.txt
# Requirements for YOLO Car Detection Training

# YOLOv8
ultralytics>=8.0.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
